restructure selectors and reducers
