+++ frontend/src/modules/app/quickstart/QuickStartPage.tsx
@@ -15,7 +15,7 @@ import {
   getParametersRequestDraftPEEP,
   getParametersRequestDraftRR,
   getParametersRequestDraftVT,
-} from '../../../store/controller/selectors';
+} from '../../../store/controller/selectors/measurements';
 import { setActiveRotaryReference } from '../Service';
 import ModeBanner, { BannerType } from '../../displays/ModeBanner';
 import { BPM, LMIN, PERCENT } from '../../info/units';
