+++ frontend/src/modules/dashboard/views/HFNCGraphView.tsx
@@ -11,11 +11,13 @@ import {
   getParametersFiO2,
   getParametersFlow,
   getROXIndex,
+} from '../../../store/controller/selectors/measurements';
+import {
   getSmoothedFiO2Value,
   getSmoothedFlow,
   getSmoothedSpO2,
   getSmoothedHR,
-} from '../../../store/controller/selectors';
+} from '../../../store/controller/selectors/derived_measurements';
 import { a11yProps, TabPanel } from '../../controllers/TabPanel';
 import { BPM, LMIN, PERCENT } from '../../info/units';
 import { FlowGraphInfo, PawGraphInfo, VolumeGraphInfo } from '../containers';
