+++ frontend/src/modules/dashboard/views/PressureControlMainView.tsx
@@ -19,7 +19,7 @@ import {
   getParametersPIP,
   getParametersRR,
   getSensorMeasurementsFiO2,
-} from '../../../store/controller/selectors';
+} from '../../../store/controller/selectors/measurements';
 import { BMIN, CMH20, PERCENT } from '../../info/units';
 import ValueInfo from '../components/ValueInfo';
