+++ frontend/src/modules/modes/ModesDropdown.tsx
@@ -12,7 +12,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { VentilationMode, ParametersRequest } from '../../store/proto/mcu_pb';
 import { MessageType } from '../../store/proto/types';
 import { commitRequest } from '../../store/controller/actions';
-import { getParametersRequestMode } from '../../store/controller/selectors';
+import { getParametersRequestMode } from '../../store/controller/selectors/measurements';
 import { getModeText } from '../displays/ModeBanner';
 
 const useStyles = makeStyles(() => ({
