+++ frontend/src/store/connection/protocols/backend/states.ts
@@ -2,14 +2,16 @@ import { OutputSelector } from 'reselect';
 import { SelectEffect, select } from 'redux-saga/effects';
 import { StoreState } from '../../../types';
 import { PBMessage, MessageType } from '../../../proto/types';
+import { getParametersRequest } from '../../../controller/selectors/measurements';
 import {
-  getParametersRequest,
   getAlarmLimitsRequest,
   getFullExpectedLogEvent,
   getAlarmMuteRequest,
+} from '../../../controller/selectors';
+import {
   getSystemSettingsRequest,
   getFrontendDisplaySetting,
-} from '../../../controller/selectors';
+} from '../../../controller/selectors/frontend_pb';
 import { getLastBackendConnectionTime } from '../../selectors';
 import {
   TaggedStateSegment,
