+++ frontend/src/store/controller/reducers.ts
@@ -30,7 +30,6 @@ import {
   eventLogReducer,
   rotaryEncoderReducer,
 } from './reducers/backend';
-import { waveformHistoryReducer, pvHistoryReducer } from './reducers/derived';
 import { REQUEST_COMMITTED, DRAFT_REQUEST_COMMITTED } from './types';
 
 export const controllerReducer = combineReducers({
@@ -77,35 +76,6 @@ export const controllerReducer = combineReducers({
     MessageType.FrontendDisplaySetting,
     REQUEST_COMMITTED,
   ),
-
-  // Derived states
-  // TODO: move plots into a separate top-level slice of the store
-  plots: combineReducers({
-    waveforms: combineReducers({
-      paw: waveformHistoryReducer<SensorMeasurements>(
-        MessageType.SensorMeasurements,
-        (sensorMeasurements: SensorMeasurements) => sensorMeasurements.time,
-        (sensorMeasurements: SensorMeasurements) => sensorMeasurements.paw,
-        60,
-        0,
-      ),
-      flow: waveformHistoryReducer<SensorMeasurements>(
-        MessageType.SensorMeasurements,
-        (sensorMeasurements: SensorMeasurements) => sensorMeasurements.time,
-        (sensorMeasurements: SensorMeasurements) => sensorMeasurements.flow,
-        60,
-        20,
-      ),
-      volume: waveformHistoryReducer<SensorMeasurements>(
-        MessageType.SensorMeasurements,
-        (sensorMeasurements: SensorMeasurements) => sensorMeasurements.time,
-        (sensorMeasurements: SensorMeasurements) => sensorMeasurements.volume,
-        60,
-        40,
-      ),
-    }),
-    pvLoop: pvHistoryReducer,
-  }),
 });
 
 export default controllerReducer;
