+++ frontend/src/store/index.ts
@@ -6,11 +6,13 @@ import { appReducer } from './app/reducers';
 import { connectionSaga } from './connection/saga';
 import { connectionReducer } from './connection/reducers';
 import { controllerReducer } from './controller/reducers';
+import { plotsReducer } from './plots/reducers';
 
 const reducers = combineReducers({
   app: appReducer,
   controller: controllerReducer,
   connection: connectionReducer,
+  plots: plotsReducer,
 });
 
 const sagaMiddleware = createSagaMiddleware();
